Fix View Attendees not triggering getSingleFlexiRecord API calls
- Add forceRefresh parameter to getVikingEventData and getVikingEventDataForEvents functions
- Pass forceRefresh=true by default for dynamic data retrieval in getVikingEventDataForEvents
- Fix parameter passing chain: getVikingEventDataForEvents → getVikingEventData → getConsolidatedFlexiRecord → getFlexiRecordData → getSingleFlexiRecord
- Ensure "View Attendees" clicks trigger fresh API calls with rate limit logging
- Fix AbortController memory leak in CampGroupsView component (CodeRabbit suggestion)
- Maintain static data caching for performance while refreshing dynamic data
- Update function signatures and JSDoc to document forceRefresh behavior
